requirements.txt
python-dotenv
python-docx
PyPDF2
pandas
openpyxl
openai>=1.0.0